+++ module/handlebars/templates.js
@@ -38,6 +38,7 @@ export const preloadHandlebarsTemplates = async function() {
     "systems/pf1/templates/apps/attack-roll-dialog.html",
     "systems/pf1/templates/apps/links.html",
     "systems/pf1/templates/apps/link-options.html",
+    "systems/pf1/templates/apps/vision-permission.html",
 
     // Chat
     "systems/pf1/templates/chat/roll-ext.html",